Refactor technology details
